Got rid of Receipt, Trip, Category and PaymentMethod interfaces
