+++ app/src/main/kotlin/co/smartreceipts/android/model/Category.kt
@@ -1,21 +1,55 @@
 package co.smartreceipts.android.model
 
 import android.os.Parcelable
-
+import co.smartreceipts.android.sync.model.SyncState
 import co.smartreceipts.android.sync.model.Syncable
+import co.smartreceipts.android.sync.model.impl.DefaultSyncState
+import kotlinx.android.parcel.Parcelize
+import java.util.*
 
-/**
- * A contract definition by which we can track receipt categories
- */
-interface Category : Parcelable, Syncable, Draggable<Category> { //TODO 18.09.18 we can get rid of interfaces
+@Parcelize
+class Category @JvmOverloads constructor(
+    override val id: Int,
+    override val uuid: UUID,
+    val name: String, // The full-name representation of this category
+    val code: String, // The "code" associated with this category
+    override val syncState: SyncState = DefaultSyncState(),
+    override val customOrderId: Long = id.toLong()
+) : Parcelable, Syncable, Draggable<Category> {
 
-    /**
-     * The full-name representation of this category
-     */
-    val name: String
+    override fun toString() = name
 
-    /**
-     * The "code" associated with this category
-     */
-    val code: String
+    override fun compareTo(other: Category): Int {
+        return if (customOrderId == 0L && other.customOrderId == 0L) {
+            name.compareTo(other.name)
+        } else {
+            customOrderId.compareTo( other.customOrderId)
+        }
+    }
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+
+        other as Category
+
+        if (id != other.id) return false
+        if (uuid != other.uuid) return false
+        if (name != other.name) return false
+        if (code != other.code) return false
+        if (customOrderId != other.customOrderId) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        var result = id
+        result = 31 * result + uuid.hashCode()
+        result = 31 * result + name.hashCode()
+        result = 31 * result + code.hashCode()
+        result = 31 * result + customOrderId.hashCode()
+        return result
+    }
+
+
 }
