+++ app/src/main/kotlin/co/smartreceipts/android/model/PaymentMethod.kt
@@ -1,19 +1,55 @@
 package co.smartreceipts.android.model
 
+import android.content.res.Resources
 import android.os.Parcelable
-
+import co.smartreceipts.android.model.factory.PaymentMethodBuilderFactory
+import co.smartreceipts.android.sync.model.SyncState
 import co.smartreceipts.android.sync.model.Syncable
+import co.smartreceipts.android.sync.model.impl.DefaultSyncState
+import kotlinx.android.parcel.Parcelize
+import java.util.*
 
 /**
- * This interface establishes the contract for encapsulating the behavior of different payment models.
+ * An immutable implementation of [PaymentMethod].
  *
  * @author Will Baumann
  */
-interface PaymentMethod : Parcelable, Syncable, Draggable<PaymentMethod> {
+@Parcelize
+class PaymentMethod @JvmOverloads constructor (
+    override val id: Int,
+    override val uuid: UUID,
+    val method: String, // The actual payment method that the user specified
+    override val syncState: SyncState = DefaultSyncState(),
+    override val customOrderId: Long = 0
 
-    /**
-     * The actual payment method that the user specified
-     */
-    val method: String
+) : Parcelable, Syncable, Draggable<PaymentMethod> {
 
+    override fun toString() = method
+
+    override fun equals(o: Any?): Boolean {
+        if (this === o) return true
+        if (o !is PaymentMethod) return false
+
+        val that = o as PaymentMethod?
+
+        if (id != that!!.id) return false
+        if (uuid != that.uuid) return false
+        return if (customOrderId != that.customOrderId) false else method == that.method
+    }
+
+    override fun hashCode(): Int {
+        var result = id
+        result = 31 * result + uuid.hashCode()
+        result = 31 * result + method.hashCode()
+        result = 31 * result + (customOrderId xor customOrderId.ushr(32)).toInt()
+        return result
+    }
+
+    override fun compareTo(paymentMethod: PaymentMethod): Int {
+        return customOrderId.compareTo(paymentMethod.customOrderId)
+    }
+
+    companion object {
+        val NONE = PaymentMethodBuilderFactory().setMethod(Resources.getSystem().getString(android.R.string.untitled)).build()
+    }
 }
