+++ app/src/main/kotlin/co/smartreceipts/android/model/Receipt.kt
@@ -2,172 +2,128 @@ package co.smartreceipts.android.model
 
 import android.content.Context
 import android.os.Parcelable
+import co.smartreceipts.android.model.utils.ModelUtils
+import co.smartreceipts.android.persistence.DatabaseHelper
+import co.smartreceipts.android.sync.model.SyncState
 import co.smartreceipts.android.sync.model.Syncable
+import kotlinx.android.parcel.Parcelize
 import java.io.File
 import java.sql.Date
 import java.util.*
 
-interface Receipt : Parcelable, Priceable, Draggable<Receipt>, Syncable {
-
+/**
+ * A mostly immutable implementation of the [Receipt] interface that
+ * serves as the default implementation.
+ */
+@Parcelize
+class Receipt constructor(
+    override val id: Int,
+    override val uuid: UUID,
     /**
-     * Gets the parent trip for this receipt. This can only be null if it's detached from a [Trip]
+     * The "index" of this receipt relative to others. If this was the second earliest receipt, it would appear
+     * as a receipt of index 2.
+     */
+    val index: Int,
+    /**
+     * The parent trip for this receipt. This can only be null if it's detached from a [Trip]
      * (e.g. if it's a converted distance).
-     *
-     * @return - the parent [Trip]
      */
-    val trip: Trip
+    val trip: Trip,
+    /**
+     * The file attached to this receipt or `null` if none is presentFirstTimeInformation
+     */
+    val file: File?,
+    /**
+     * The payment method associated with this receipt item.
+     */
+    val paymentMethod: PaymentMethod,
+    /**
+     * The name of this receipt. This should never be `null`.
+     */
+    val name: String,
+    /**
+     * The [Category] to which this receipt is attached
+     */
+    val category: Category,
+    /**
+     * The user defined comment for this receipt
+     */
+    val comment: String,
+    override val price: Price,
+    /**
+     * The [Price] tax associated with this receipt
+     */
+    val tax: Price,
+    /**
+     * The [Date] during which this receipt was taken
+     */
+    val date: Date,
+    /**
+     * The [TimeZone] in which the date was set
+     */
+    val timeZone: TimeZone,
+    /**
+     * Checks if the receipt was marked as Reimbursable (i.e. counting towards the total) or not
+     */
+    val isReimbursable: Boolean,
+    /**
+     * Checks if this receipt should be printed as a full page in the PDF report
+     */
+    val isFullPage: Boolean,
+    /**
+     *  Checks if this receipt is currently selected or not
+     */
+    val isSelected: Boolean,
+    /**
+     * The [Source] from which this receipt was built for debugging purposes
+     */
+    val source: Source,
+    private val extraEditTextOne: String?,
+    private val extraEditTextTwo: String?,
+    private val extraEditTextThree: String?,
+    override val syncState: SyncState,
+    override val customOrderId: Long
+    ) : Parcelable, Priceable, Draggable<Receipt>, Syncable {
+
+    val extraEditText1: String? = if (DatabaseHelper.NO_DATA == extraEditTextOne) null else extraEditTextOne
+    val extraEditText2: String? = if (DatabaseHelper.NO_DATA == extraEditTextTwo) null else extraEditTextTwo
+    val extraEditText3: String? = if (DatabaseHelper.NO_DATA == extraEditTextThree) null else extraEditTextThree
 
     /**
-     * Gets the payment method associated with this receipt item.
-     *
-     * @return the [co.smartreceipts.android.model.PaymentMethod] associated with this receipt item.
-     */
-    val paymentMethod: PaymentMethod
-
-    /**
-     * Gets the name of this receipt. This should never be `null`.
-     *
-     * @return the name of this receipt as a [String].
-     */
-    val name: String
-
-    /**
-     * Gets the file attached to this receipt.
-     *
-     * @return the Image [File] or `null` if none is presentFirstTimeInformation
-     */
-    val file: File?
-
-    /**
-     * Gets the absolute path of this Receipt's file from [.getFile].
-     *
-     * @return a representation of the file path via [.getFile] and [File.getAbsolutePath].
-     */
-    val filePath: String
-
-    /**
-     * Gets the name of this Receipt's file from [.getFile].
-     *
-     * @return a representation of the file name via [.getFile] and [File.getName].
+     * The name of this Receipt's file from [.getFile].
      */
     val fileName: String
+        get() = file?.name ?: ""
 
     /**
+     * The last updated time or `-1` if we don't have a file
+     *
      * Java uses immutable [File], so when we rename our files as part of a receipt update, we might rename it
      * to the same file name as the old receipt. By tracking the last update time as well, we can determine if this file
      * was updated between two "like" receipts
-     *
-     * @return the last updated time or `-1` if we don't have a file
      */
     val fileLastModifiedTime: Long
+        get() = file?.lastModified() ?: -1
 
     /**
-     * Gets the source from which this receipt was built for debugging purposes
-     *
-     * @return the [Source]
+     * The absolute path of this Receipt's file from [.getFile].
      */
-    val source: Source
+    val filePath: String
+        get() = file?.absolutePath ?: ""
 
     /**
-     * Gets the category to which this receipt is attached
-     *
-     * @return the [Category] this receipt uses
+     *  Checks if this receipt is connected to an image file
      */
-    val category: Category
+    fun hasImage(): Boolean {
+        return file?.name?.run { endsWith(".jpg") || endsWith(".jpeg") || endsWith(".png") } ?: false
+    }
 
     /**
-     * Gets the user defined comment for this receipt
-     *
-     * @return - the current comment as a [String]
+     *  Checks if this receipt is connected to an PDF file
      */
-    val comment: String
-
-
-    /**
-     * Gets the tax associated with this receipt
-     *
-     * @return the [Price] for the tax
-     */
-    val tax: Price
-
-
-    /**
-     * Returns the date during which this receipt was taken
-     *
-     * @return the [Date] this receipt was captured
-     */
-    val date: Date
-
-    /**
-     * Gets the time zone in which the date was set
-     *
-     * @return - the [TimeZone] for the date
-     */
-    val timeZone: TimeZone
-
-    /**
-     * Checks if the receipt was marked as Reimbursable (i.e. counting towards the total) or not
-     *
-     * @return `true` if it's Reimbursable, `false` otherwise
-     */
-    val isReimbursable: Boolean
-
-    /**
-     * Checks if this receipt should be printed as a full page in the PDF report
-     *
-     * @return `true` if it's printed as a full page, `false` otherwise
-     */
-    val isFullPage: Boolean
-
-    /**
-     * Checks if this receipt is currently selected or not
-     *
-     * @return `true` if it's currently selected. `false` otherwise
-     */
-    val isSelected: Boolean
-
-    /**
-     * Returns the "index" of this receipt relative to others. If this was the second earliest receipt, it would appear
-     * as a receipt of index 2.
-     *
-     * @return the index of this receipt
-     */
-    val index: Int
-
-    /**
-     * Returns the user defined string for the 1st "extra" field
-     *
-     * @return the [String] for the 1st custom field or `null` if not set
-     */
-    val extraEditText1: String?
-
-    /**
-     * Returns the user defined string for the 2nd "extra" field
-     *
-     * @return the [String] for the 2nd custom field or `null` if not set
-     */
-    val extraEditText2: String?
-
-    /**
-     * Returns the user defined string for the 3rd "extra" field
-     *
-     * @return the [String] for the 3rd custom field or `null` if not set
-     */
-    val extraEditText3: String?
-
-    /**
-     * Checks if this receipt is connected to an image file
-     *
-     * @return `true` if it has an image file, `false` otherwise
-     */
-    fun hasImage(): Boolean
-
-    /**
-     * Checks if this receipt is connected to an PDF file
-     *
-     * @return `true` if it has a PDF file, `false` otherwise
-     */
-    fun hasPDF(): Boolean
+    fun hasPDF(): Boolean {
+        return file?.name?.endsWith(".pdf") ?: false
+    }
 
     /**
      * Gets a formatted version of the date based on the timezone and locale for a given separator. In the US,
@@ -177,31 +133,108 @@ interface Receipt : Parcelable, Priceable, Draggable<Receipt>, Syncable {
      * @param separator - the date separator (e.g. "/", "-", ".")
      * @return the formatted date string for this receipt
      */
-    fun getFormattedDate(context: Context, separator: String): String
+    fun getFormattedDate(context: Context, separator: String): String {
+        return ModelUtils.getFormattedDate(date, timeZone, context, separator)
+    }
 
-    /**
-     * Checks if we have a 1st "extra" field
-     *
-     * @return `true` if we have a 1st "extra" field or `false` if not
-     */
-    fun hasExtraEditText1(): Boolean
+    fun hasExtraEditText1(): Boolean = extraEditText1 != null
 
-    /**
-     * Checks if we have a 2nd "extra" field
-     *
-     * @return `true` if we have a 2nd "extra" field or `false` if not
-     */
-    fun hasExtraEditText2(): Boolean
+    fun hasExtraEditText2(): Boolean = extraEditText2 != null
 
-    /**
-     * Checks if we have a 3rd "extra" field
-     *
-     * @return `true` if we have a 3rd "extra" field or `false` if not
-     */
-    fun hasExtraEditText3(): Boolean
+    fun hasExtraEditText3(): Boolean = extraEditText3 != null
+
+    override fun toString(): String {
+        return "DefaultReceiptImpl{" +
+                "id=" + id +
+                ", uuid='" + uuid.toString() +
+                ", name='" + name + '\''.toString() +
+                ", trip=" + trip.name +
+                ", paymentMethod=" + paymentMethod +
+                ", index=" + index +
+                ", comment='" + comment + '\''.toString() +
+                ", category=" + category +
+                ", price=" + price.currencyFormattedPrice +
+                ", tax=" + tax +
+                ", date=" + date +
+                ", timeZone=" + timeZone.id +
+                ", isReimbursable=" + isReimbursable +
+                ", isFullPage=" + isFullPage +
+                ", source=" + source +
+                ", extraEditText1='" + extraEditText1 + '\''.toString() +
+                ", extraEditText2='" + extraEditText2 + '\''.toString() +
+                ", extraEditText3='" + extraEditText3 + '\''.toString() +
+                ", isSelected=" + isSelected +
+                ", file=" + file +
+                ", fileLastModifiedTime=" + fileLastModifiedTime +
+                ", customOrderId=" + customOrderId +
+                '}'.toString()
+    }
+
+    override fun equals(o: Any?): Boolean {
+        if (this === o) return true
+        if (o !is Receipt) return false
+
+        val that = o as Receipt?
+
+        if (id != that!!.id) return false
+        if (uuid != that.uuid) return false
+        if (isReimbursable != that.isReimbursable) return false
+        if (isFullPage != that.isFullPage) return false
+        if (trip != that.trip) return false
+        if (paymentMethod != that.paymentMethod) return false
+        if (index != that.index) return false
+        if (name != that.name) return false
+        if (comment != that.comment) return false
+        if (category != that.category) return false
+        if (price != that.price) return false
+        if (tax != that.tax) return false
+        if (date != that.date) return false
+        if (timeZone != that.timeZone) return false
+        if (if (extraEditText1 != null) extraEditText1 != that.extraEditText1 else that.extraEditText1 != null)
+            return false
+        if (if (extraEditText2 != null) extraEditText2 != that.extraEditText2 else that.extraEditText2 != null)
+            return false
+        if (if (extraEditText3 != null) extraEditText3 != that.extraEditText3 else that.extraEditText3 != null)
+            return false
+        if (fileLastModifiedTime != that.fileLastModifiedTime) return false
+        if (customOrderId != that.customOrderId) return false
+        return if (file != null) file == that.file else that.file == null
+
+    }
+
+    override fun hashCode(): Int {
+        var result = id
+        result = 31 * result + uuid.hashCode()
+        result = 31 * result + trip.hashCode()
+        result = 31 * result + paymentMethod.hashCode()
+        result = 31 * result + index
+        result = 31 * result + name.hashCode()
+        result = 31 * result + comment.hashCode()
+        result = 31 * result + category.hashCode()
+        result = 31 * result + price.hashCode()
+        result = 31 * result + tax.hashCode()
+        result = 31 * result + date.hashCode()
+        result = 31 * result + timeZone.hashCode()
+        result = 31 * result + if (isReimbursable) 1 else 0
+        result = 31 * result + if (isFullPage) 1 else 0
+        result = 31 * result + (extraEditText1?.hashCode() ?: 0)
+        result = 31 * result + (extraEditText2?.hashCode() ?: 0)
+        result = 31 * result + (extraEditText3?.hashCode() ?: 0)
+        result = 31 * result + if (file != null) file.hashCode() else 0
+        result = 31 * result + fileLastModifiedTime.toInt()
+        result = 31 * result + (customOrderId xor customOrderId.ushr(32)).toInt()
+        return result
+    }
+
+    override fun compareTo(receipt: Receipt): Int {
+        return if (customOrderId == receipt.customOrderId) {
+            receipt.date.compareTo(date)
+        } else {
+            -java.lang.Long.compare(customOrderId, receipt.customOrderId)
+        }
+    }
 
     companion object {
-
         val PARCEL_KEY: String = Receipt::class.java.name
     }
